Fixed issue with SNPP data when running static microsimulation
When running the static microsimulation (year by year), each year uses a data source to constrain the microsynthesis. When moving from 2017 to 2018 (projection changed from SNPP (2016 based) to NPP (2018 based)), the split_range() function was returning a list with an empty element, instead of a None type object. This was causing an error in SNPPData.filter() where the normal logic for handling a None type object was not being executed due to the list being real but empty. Now the None type list is converted to None before being passed to filter().